// src/types/generated/quiz.pkl.ts
/* This file was generated by `pkl-typescript` from Pkl module `Quiz`. */
/* DO NOT EDIT! */
/* istanbul ignore file */
/* eslint-disable */

// Ref: Module root.
export interface Quiz {
  title: string

  categories: Array<Category>
}

// Ref: Pkl class `Quiz.Category`.
export interface Category {
  slug: string

  name: string

  displayName: string

  options: Array<Option>
}

// Ref: Pkl class `Quiz.Option`.
export interface Option {
  name: string

  difficulty: number
}